# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VerifyMzansi API",
    "description": "Public and internal API for South Africa's verification-first marketplace.",
    "version": "1.0.0",
    "contact": {
      "name": "VerifyMzansi",
      "url": "https://verifymzansi.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "https://verifymzansi.com", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/verification/session/start": {
      "post": {
        "tags": ["Verification"],
        "summary": "Start a KYC verification session",
        "operationId": "startVerificationSession",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/verification/status": {
      "get": {
        "tags": ["Verification"],
        "summary": "Get current KYC verification status",
        "operationId": "getVerificationStatus",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Verification status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerificationStatus" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie (set automatically on login)"
      }
    },
...
    "schemas": {
...
      "VerificationStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["none", "pending", "approved", "rejected"]
          },
          "submittedAt": { "type": "string", "format": "date-time" },
          "decidedAt": { "type": "string", "format": "date-time" },
          "rejectionReason": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  }
}