# src/app/api/verification/session/start/route.ts
/**
 * POST /api/verification/session/start
 * Starts or resumes a verification session and returns step requirements.
 */

import { type NextRequest, NextResponse } from "next/server";
import { isFeatureEnabled } from "@/lib/services/feature-flags";
import { logAuditEvent } from "@/lib/services/audit";
import { REQUIRED_VERIFICATION_STEPS } from "@/lib/constants/verification";
import { createLogger } from "@/lib/utils/logger";
import { createAdminClient } from "@/lib/supabase/admin";
import { checkRateLimit } from "@/lib/utils/rate-limit";
import { rateLimitExceededResponse } from "@/lib/utils/rate-limit-responses";
import { enforceConfirmedVerificationRequest } from "../../_lib/verification-request-prelude";

const log = createLogger("SessionStart");

export async function POST(_request: NextRequest) {
  try {
    const request = _request;
    const prelude = await enforceConfirmedVerificationRequest(request, log);
    if (!prelude.success) return prelude.response;

    const { supabase, user } = prelude;

    const rateCheck = await checkRateLimit({
      key: user.id,
      action: "verification:session-start",
      degradedMode: "block",
    });
    if (rateCheck.limited) {
      return rateLimitExceededResponse({
        degraded: rateCheck.degraded,
        retryAfter: rateCheck.retryAfter,
        degradedMessage:
          "Verification session protection is temporarily unavailable. Please try again.",
        limitedMessage: "Too many verification session attempts. Please try again later.",
      });
    }

    // Feature flag check
    const v2Enabled = await isFeatureEnabled("kyc_v2_flow");
    if (!v2Enabled) {
      return NextResponse.json(
        { error: "New verification flow is not yet enabled" },
        { status: 404 }
      );
    }

    // Fetch the most recent verification session, including finalized ones.
    // Finalized sessions must stay finalized on revisit so the UI can keep the
    // user in the completed state until a real resubmission reopens the session.
    const { data: existingSession, error: sessionFetchErr } = await supabase
      .from("verification_sessions")
      .select("*")
      .eq("user_id", user.id)
      .order("created_at", { ascending: false })
      .limit(1)
      .maybeSingle();

    if (sessionFetchErr) {
      log.error("Failed to fetch existing verification session", {
        userId: user.id,
        error: sessionFetchErr.message,
      });
      return NextResponse.json({ error: "Failed to load verification session" }, { status: 500 });
    }

    let session = existingSession;

    // If the existing session is expired, reset it for reuse (UNIQUE(user_id) allows only one row)
    if (session && !session.finalized_at) {
      const expiresAt = new Date(new Date(session.created_at).getTime() + 24 * 60 * 60 * 1000);
      if (expiresAt < new Date()) {
        // Check which steps are already submitted to preserve their state.
        // Pending identity/location steps are still valid user submissions while admin reviews them.
        const { data: submittedSteps, error: submittedStepsErr } = await supabase
          .from("verification_steps")
          .select("step_type, status, phone_verified_at")
          .eq("user_id", user.id)
          .in("status", ["approved", "pending"]);

        if (submittedStepsErr) {
          log.warn("Failed to fetch submitted steps during session reset (non-fatal)", {
            userId: user.id,
            error: submittedStepsErr.message,
          });
        }

        const submittedTypes = new Set((submittedSteps || []).map((s) => s.step_type));
        const phoneStep = (submittedSteps || []).find(
          (s) => s.step_type === "phone" && s.phone_verified_at
        );

        // Reset the expired session in-place instead of finalize + insert
        // (inserting would violate the UNIQUE(user_id) constraint)
        // Preserve artifact references for already-approved steps.
        // Writes go through the service-role admin client: the owner UPDATE
        // policy on verification_sessions was dropped and finalized_at /
        // phone_verified_at are guard-trigger protected signal columns.
        const admin = createAdminClient();
        const { data: resetSession, error: resetErr } = await admin
          .from("verification_sessions")
          .update({
            finalized_at: null,
            created_at: new Date().toISOString(),
            phone_verified_at: phoneStep?.phone_verified_at ?? null,
            id_artifact_id: submittedTypes.has("id_doc") ? session.id_artifact_id : null,
            selfie_artifact_id: submittedTypes.has("selfie") ? session.selfie_artifact_id : null,
            location_submitted_at: submittedTypes.has("location")
              ? session.location_submitted_at
              : null,
          })
          .eq("id", session.id)
          .select()
          .single();

        if (resetErr || !resetSession) {
          log.error("Failed to reset expired session", { error: resetErr?.message ?? "unknown" });
          return NextResponse.json(
            { error: "Failed to create verification session" },
            { status: 500 }
          );
        }

        session = resetSession;

        await logAuditEvent({
          actorId: user.id,
          actorRole: "member",
          action: "kyc_session_started",
          targetType: "verification_session",
          targetId: session.id,
        });
      }
    }

    if (!session) {
      // Check if phone was already verified in verification_steps
      const { data: phoneStep, error: phoneStepErr } = await supabase
        .from("verification_steps")
        .select("phone_verified_at")
        .eq("user_id", user.id)
        .eq("step_type", "phone")
        .in("status", ["approved", "pending"])
        .maybeSingle();

      if (phoneStepErr) {
        log.warn("Failed to fetch phone step (non-fatal)", {
          userId: user.id,
          error: phoneStepErr.message,
        });
      }

      // Use upsert to handle edge case where a finalized row already exists.
      // Written via the service-role admin client (see note above): inserts
      // carry phone_verified_at, which only verification workflows may set.
      const admin = createAdminClient();
      const { data: newSession, error: insertErr } = await admin
        .from("verification_sessions")
        .upsert(
          {
            user_id: user.id,
            finalized_at: null,
            phone_verified_at: phoneStep?.phone_verified_at ?? null,
            id_artifact_id: null,
            selfie_artifact_id: null,
            location_submitted_at: null,
          },
          { onConflict: "user_id" }
        )
        .select()
        .single();

      if (insertErr || !newSession) {
        log.error("Failed to create session", { error: insertErr?.message ?? "unknown" });
        return NextResponse.json(
          { error: "Failed to create verification session" },
          { status: 500 }
        );
      }

      session = newSession;

      await logAuditEvent({
        actorId: user.id,
        actorRole: "member",
        action: "kyc_session_started",
        targetType: "verification_session",
        targetId: session.id,
      });
    }

    // Fetch all existing verification steps for this user
    const { data: steps, error: stepsErr } = await supabase
      .from("verification_steps")
      .select("step_type, status")
      .eq("user_id", user.id);

    if (stepsErr) {
      log.warn("Failed to fetch verification steps (non-fatal)", {
        userId: user.id,
        error: stepsErr.message,
      });
    }

    const completedSteps: string[] = [];
    const pendingSteps: string[] = [];
    const rejectedSteps: string[] = [];

    for (const step of steps || []) {
      if (step.status === "approved") {
        completedSteps.push(step.step_type);
      } else if (step.status === "pending") {
        pendingSteps.push(step.step_type);
      } else if (step.status === "rejected" || step.status === "needs_resubmission") {
        rejectedSteps.push(step.step_type);
      }
    }

    // Calculate session expiry (24h from start)
    const expiresAt = new Date(
      new Date(session.created_at).getTime() + 24 * 60 * 60 * 1000
    ).toISOString();

    return NextResponse.json({
      sessionId: session.id,
      completedSteps,
      pendingSteps,
      rejectedSteps,
      requiredSteps: [...REQUIRED_VERIFICATION_STEPS],
      expiresAt,
      phoneVerifiedAt: session.phone_verified_at,
      idArtifactId: session.id_artifact_id,
      selfieArtifactId: session.selfie_artifact_id,
      locationSubmittedAt: session.location_submitted_at,
      finalizedAt: session.finalized_at,
    });
  } catch (err) {
    log.error("Unexpected error", {
      error: err instanceof Error ? err.message : "unknown error",
      stack: err instanceof Error ? err.stack : undefined,
    });
    return NextResponse.json({ error: "Internal server error" }, { status: 500 });
  }
}